refactor: retirnado botão de logout da home #44
